Move tests to their own file
